Refactor creation of ExpressionFactory
ELUtil doesn't require a factory for all operations whereas the
ELManager does. Moving construction of the factory to the ELManager
allows some limited unit testing of ELUtil without placing additional
limits on how ELManager can be used.
