+++ api/src/main/java/jakarta/el/ELContext.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1997, 2021 Oracle and/or its affiliates and others.
+ * Copyright (c) 1997, 2022 Oracle and/or its affiliates and others.
  * All rights reserved.
  * Copyright 2004 The Apache Software Foundation
  *
@@ -436,7 +436,8 @@ public abstract class ELContext {
         } finally {
             setPropertyResolved(propertyResolvedSave);
         }
-        return ELUtil.getExpressionFactory().coerceToType(obj, targetType);
+
+        return ELManager.getExpressionFactory().coerceToType(obj, targetType);
     }
 
 }
